fix: set repo url also for release command
