+++ crates/release_plz/src/args/mod.rs
@@ -2,6 +2,7 @@ mod generate_completions;
 mod release;
 mod release_pr;
 mod update;
+mod repo_command;
 
 use std::path::{Path, PathBuf};
